move relay::v2::client symbols to relay::client and use them across the crate
