+++ protocols/relay/src/client/handler.rs
@@ -18,9 +18,9 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
+use crate::client::transport;
 use crate::message_proto::Status;
 use crate::protocol::{self, inbound_stop, outbound_hop};
-use crate::v2::client::transport;
 use either::Either;
 use futures::channel::{mpsc, oneshot};
 use futures::future::{BoxFuture, FutureExt};
