+++ protocols/relay/src/lib.rs
@@ -23,6 +23,7 @@
 
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+pub mod client;
 mod protocol;
 pub mod v2;
