+++ protocols/relay/tests/v2.rs
@@ -31,7 +31,7 @@ use libp2p_core::PublicKey;
 use libp2p_core::{identity, upgrade, PeerId};
 use libp2p_ping as ping;
 use libp2p_plaintext::PlainText2Config;
-use libp2p_relay::v2::client;
+use libp2p_relay::client;
 use libp2p_relay::v2::relay;
 use libp2p_swarm::{AddressScore, NetworkBehaviour, Swarm, SwarmEvent};
 use std::time::Duration;
